fix logic update profile

+++ src/main/java/com/example/social_network_api/service/impl/UserServiceImpl.java
@@ -11,12 +11,10 @@ import com.example.social_network_api.exception.custom.ConflictException;
 import com.example.social_network_api.exception.custom.ForbiddenException;
 import com.example.social_network_api.exception.custom.ResourceNotFoundException;
 import com.example.social_network_api.repository.PasswordResetTokenRepositoty;
+import com.example.social_network_api.repository.ProfileRepository;
 import com.example.social_network_api.repository.UserRepository;
 import com.example.social_network_api.security.jwt.JWTUtils;
-import com.example.social_network_api.service.TokenStoreService;
-import com.example.social_network_api.service.MailService;
-import com.example.social_network_api.service.RoleService;
-import com.example.social_network_api.service.UserService;
+import com.example.social_network_api.service.*;
 import com.example.social_network_api.utils.AuthUtils;
 import io.jsonwebtoken.Claims;
 import jakarta.transaction.Transactional;
@@ -48,6 +46,7 @@ import java.util.stream.Collectors;
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final ProfileRepository profileRepository;
     private final PasswordEncoder passwordEncoder;
     private final AuthenticationManager authenticationManager;
     private final JWTUtils jwtUtils;
@@ -304,6 +303,10 @@ public class UserServiceImpl implements UserService {
     public void deleteById(Long id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found"));
+        if(!AuthUtils.isAdmin()){
+            throw new ForbiddenException("Unauthorized.");
+        }
+        profileRepository.deleteByUserId(user.getId());
         userRepository.deleteById(user.getId());
     }
 
@@ -312,6 +315,9 @@ public class UserServiceImpl implements UserService {
     public void disableUser(Long id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found"));
+        if(!AuthUtils.getCurrentUsername().equals(user.getUsername()) && !AuthUtils.isAdmin()){
+            throw new ForbiddenException("Unauthorized.");
+        }
         user.setEnabled(false);
         userRepository.save(user);
     }
@@ -321,6 +327,9 @@ public class UserServiceImpl implements UserService {
     public void enableUser(Long id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found"));
+        if(!AuthUtils.getCurrentUsername().equals(user.getUsername()) && !AuthUtils.isAdmin()){
+            throw new ForbiddenException("Unauthorized.");
+        }
         user.setEnabled(true);
         userRepository.save(user);
     }
